third_party/blink/web_tests/virtual/stylable-select-disabled/html5lib/generated/run-tests17-data-expected.txt

CONSOLE WARNING: A tr tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A td tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A th tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A tbody tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A thead tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A tfoot tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A caption tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
../resources/tests17.dat: PASS
